Avoid crash in `/list` with broken servers
Some servers appear to not send `RPL_LISTSTART` when starting returning the channel list, which caused a crash due to the array not existing. This adds both a client method to explicitely clear the list, and also adds a safeguard around the array in case or a surprise list response from the server (for example, another client over ZNC).