feat: supports CLIENTTAGDENY parsing and "is blocked" function
This introduces a `MessageTags.parseDenylist` which parses
the value of `CLIENTTAGDENY=<value>` string into a structure containing
everything you need to perform O(#allowed tags) or O(#denied tags)
check for whether a tag is allowed or not.

Then, a `MessageTags.isBlocked` function can be used to leverage that
structure to perform this check.